fix: OrderForm items 타입을 Integer로 수정

+++ src/main/java/com/test08/domain/order/dto/OrderForm.java
@@ -6,5 +6,6 @@ public record OrderForm(
         String email,
         String address,
         String postCode,
-        Map<Long, Integer> items
-) {}
+        Map<Integer, Integer> items
+) {
+}
